+++ src/views/default/update.php
@@ -29,6 +29,17 @@ use vova07\imperavi\Widget as ImperaviRedactor;
 <div class="container">
     <div class="row">
         <div class="col-md-12">
+            <h1>
+                <?= TemplatesModule::t('Email templates') ?>
+                <small><?= TemplatesModule::t('update template') ?></small>
+                <div class="pull-right">
+                    <?= LanguagesList::widget(['currentLanguage' => $translation->language]) ?>
+                </div>
+            </h1>
+        </div>
+        <div class="clearfix"></div>
+        <hr>
+        <div class="col-md-12">
             <?php if (isset($errors)): ?>
                 <?php foreach ($errors as $fieldErrors): ?>
                     <?php foreach ($fieldErrors as $error): ?>
@@ -41,11 +52,6 @@ use vova07\imperavi\Widget as ImperaviRedactor;
                     <?php endforeach; ?>
                 <?php endforeach; ?>
             <?php endif; ?>
-        </div>
-        <div class="col-md-12">
-            <div class="pull-right">
-                <?= LanguagesList::widget(['currentLanguage' => $translation->language]) ?>
-            </div>
         </div>
         <div class="col-md-12">
             <?php $form = ActiveForm::begin() ?>
@@ -70,5 +76,6 @@ use vova07\imperavi\Widget as ImperaviRedactor;
             ) ?>
             <?php $form->end() ?>
         </div>
+        <?= $this->render('_issue-message') ?>
     </div>
 </div>